% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_set_enrichment.R
\name{calc_enrichment}
\alias{calc_enrichment}
\title{Calculates enrichment p-values based on GREGOR}
\usage{
calc_enrichment(overlap)
}
\arguments{
\item{overlap}{The \code{data.table} returned by \code{overlap}}
}
\value{
A enrichment p-value
}
\description{
Calculates enrichment p-values based on GREGOR
}
