% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_set_enrichment.R
\name{calc_odds_ratio}
\alias{calc_odds_ratio}
\title{Calculates odds ratio between foreground variant overlap and background 
variant overlap (with the genomic region)}
\usage{
calc_odds_ratio(overlap)
}
\arguments{
\item{overlap}{The \code{data.table} returned by overlap}
}
\value{
A \code{data.table} with mean and se of the odds ratio
}
\description{
Calculates odds ratio between foreground variant overlap and background 
variant overlap (with the genomic region)
}
