% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_set_enrichment.R
\name{count_overlap}
\alias{count_overlap}
\title{Count overlap between variant and region}
\usage{
count_overlap(overlap)
}
\arguments{
\item{overlap}{The object returned by overlap}
}
\value{
The number of overlap
}
\description{
Count overlap between variant and region
}
