% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_background_variants.R
\name{create_index_set}
\alias{create_index_set}
\title{Merge the background_ls into an index_set}
\usage{
create_index_set(background_ls, snpsnap)
}
\arguments{
\item{background_ls}{A list returned by \code{select_background_variants}}

\item{snpsnap}{The SNPsnap dataset returned by \code{read_snpsnap}}
}
\value{
A \code{data.table} object
}
\description{
Merge the background_ls into an index_set
}
\details{
The returned \code{data.table} contains both the foreground variant 
and matched background variants. Each background variant is marked with the 
respective foreground variant in the \code{foreground_variant} column.
}
