% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_regions}
\alias{make_regions}
\title{Helper function to make a region \code{data.table}}
\usage{
make_regions(index_set, prob_fg, prob_bg, seed = 42)
}
\arguments{
\item{index_set}{The \code{data.table} returned by \code{create_index_set}}

\item{prob}{The probability of sampling a region that contains a variant in
ld_set}
}
\description{
Helper function to make a region \code{data.table}
}
