% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_set_enrichment.R
\name{overlap}
\alias{overlap}
\title{Find overlap between variants and genomic region}
\usage{
overlap(ld_set, region)
}
\arguments{
\item{ld_set}{The \code{data.table} returned by \code{select_LD_variants}}

\item{region}{A \code{data.table} of genomic regions. Must have three columns 
chr, start and end}
}
\value{
A \code{data.table} that contains overlap between variants and 
genomic regions
}
\description{
Find overlap between variants and genomic region
}
