% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_background_variants.R
\name{read_snpsnap}
\alias{read_snpsnap}
\title{Load the SNPsnap dataset (https://data.broadinstitute.org/mpg/snpsnap/). If the dataset does not exist on the local computer, it will be downloaded. The file size is ~1G.}
\usage{
read_snpsnap(population = c("EUR", "EAS", "WAFR"))
}
\arguments{
\item{population}{the population of your study}
}
\value{
the SNPsnap dataset
}
\description{
Load the SNPsnap dataset (https://data.broadinstitute.org/mpg/snpsnap/). If the dataset does not exist on the local computer, it will be downloaded. The file size is ~1G.
}
