% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_background_variants.R
\name{select_background_variants}
\alias{select_background_variants}
\title{Select matched background variants}
\usage{
select_background_variants(x, snpsnap, bg_size = 200)
}
\arguments{
\item{x}{A vector of SNP IDs (format: chr:pos, e.g. 12:56115585)}

\item{snpsnap}{The SNPsnap dataset returned by \code{read_snpsnap}}

\item{bg_size}{The number of background variant per SNP ID. Default = 200}
}
\value{
A list of \code{data.table}s. Each \code{data.table} contains the 
background variant for a SNP ID.
}
\description{
Select matched background variants
}
\details{
For each SNP ID, \code{select_background_variants} select 
\code{bg_size} number of background variants matched by minor allele 
frequency, distance to the nearest protein coding gene, and the number 
of LD proxies (r2>0.7).
}
