% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_LD_variants.R
\name{select_LD_variants}
\alias{select_LD_variants}
\title{Select variants in LD with the index variants}
\usage{
select_LD_variants(index_set, population, plink)
}
\arguments{
\item{index_set}{The output from \code{create_index_set}}

\item{population}{The population is one of AFR, EUR or EAS}

\item{plink}{The path to plink}
}
\value{
A \code{data.table} that contains the original variants in index_set, 
as well as their LD proxies (r2>0.7).
}
\description{
Select variants in LD with the index variants
}
\details{
\code{select_LD_variants} selects LD proxies for variants in the 
index_set based on 1000 Genomes Phase 3. The function automatically downloads
1000 Genomces Phase 3 VCF files for the specified population. The function 
also requires plink2 (https://www.cog-genomics.org/plink2) to perform r2 
calculation.
}
