\name{vsea-package}
\alias{vsea-package}
\alias{vsea}
\docType{package}
\title{
\packageTitle{vsea}
}
\description{
\packageDescription{vsea}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{vsea}
\packageIndices{vsea}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{vsea}

Maintainer: \packageMaintainer{vsea}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
