% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsea.R
\name{vsea}
\alias{vsea}
\title{Variant Set Enrichment Analysis}
\usage{
vsea(snp, region, population,
  plink = "/Users/boshliu/Documents/tools/plink_mac/plink", bg_size = 200)
}
\arguments{
\item{snp}{A vector of SNP IDs (format: chr:pos, e.g. 12:56115585)}

\item{region}{A \code{data.table} of genomic regions. Must have three columns 
chr, start and end}

\item{plink}{The path to plink}

\item{bg_size}{The number of background variant per SNP ID. Default = 200}
}
\value{
A list with two elements. The first element contains enrichment 
P-value. The second element contains enrichment odds ratio and standard error
}
\description{
Variant Set Enrichment Analysis
}
